+++ tests/python/pytest_ops.py
@@ -17,10 +17,7 @@ from typing import Callable
 
 from nvfuser import FusionCache, FusionDefinition
 
-
-def is_pre_volta():
-    prop = torch.cuda.get_device_properties(torch.cuda.current_device())
-    return prop.major < 7
+from utils import is_pre_volta
 
 
 def parse_args_fusion_execution(opinfo: OpInfo, *args):
